arena/backend/pyproject.toml
[project]
name = "champion"
version = "0.1.0"
description = ""
authors = [
    { name = "lyp1noff", email = "lyp1noff@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dynamic = ["dependencies"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.23.8"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.0"
httpx = "^0.28.1"
flake8 = "^7.3.0"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
champion-domain = { path = "../../domain", develop = true }
fastapi = ">=0.118.0,<0.119.0"
uvicorn = ">=0.37.0,<0.38.0"
asyncpg = ">=0.30.0,<0.31.0"
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.16.4,<2.0.0"
pydantic = ">=2.11.7,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
python-multipart = ">=0.0.20,<0.0.21"
passlib = ">=1.7.4,<2.0.0"
python-jose = ">=3.5.0,<4.0.0"
pillow = ">=11.3.0,<12.0.0"
cairosvg = ">=2.8.2,<3.0.0"
cairocffi = ">=1.7.1,<2.0.0"
cssselect2 = ">=0.8.0,<0.9.0"
defusedxml = ">=0.7.1,<0.8.0"
tinycss2 = ">=1.4.0,<2.0.0"
webencodings = ">=0.5.1,<0.6.0"
boto3 = ">=1.39.14,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
argon2-cffi = ">=25.1.0,<26.0.0"
websockets = ">=15.0.1,<16.0.0"
typer = ">=0.19.2,<0.20.0"
pypdf = ">=6.1.1,<7.0.0"
broadcaster = ">=0.3.1,<0.4.0"
redis = ">=6.4.0,<7.0.0"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
show_error_codes = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
exclude = ["tests", "scripts", "alembic"]
